backend/requirements-test.txt
# Test requirements for Gladiator Arena backend

# Core testing framework
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0

# HTTP testing for FastAPI
httpx==0.25.2

# Database testing
pytest-postgresql==5.1.1
sqlalchemy-utils==0.41.1

# Mocking and utilities
pytest-mock==3.12.0
factory-boy==3.3.0
faker==20.1.0

# Coverage reporting
coverage==7.3.2

# Performance testing
pytest-benchmark==4.0.0

# Async testing support
anyio==4.0.0

# Type checking support
mypy==1.7.1
